Refactor of Post to generate random content in constructor
